Fix keyword search

+++ api/main/util/dto.py
@@ -17,7 +17,7 @@ class KeywordDto:
     )
     keyword_count = api.model(
         "keyword",
-        {"keyword": fields.Nested(keyword), "count": fields.Float(required=True)},
+        {"keyword": fields.Nested(keyword), "count": fields.Integer(required=True)},
     )
